Moved api to hooks folder
